server/requirements.txt
opencv-python-headless==4.8.1.78
fer==22.5.1
Pillow==10.2.0
numpy==1.26.4
tensorflow-cpu==2.15.0
mtcnn==0.1.1
python-multipart==0.0.6